requirements.txt
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu126
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu126
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu126
torchmetrics==1.6.2
torchsampler==0.1.2
lightning==2.5.0.post0

scikit-learn==1.6.1

hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

onnxruntime==1.21.0
onnx==1.17.0

pyrootutils==1.0.4
mlflow==2.20.3
rich==13.9.4
openpyxl==3.1.5
protobuf==5.29.3
tqdm==4.67.1
pillow==11.0.0
numpy==2.2.2
pandas==2.2.3
matplotlib==3.10.0
tensorboard==2.19.0
